requirements.txt
aiogram==3.4.0
Pillow==10.4.0
pikepdf==9.2.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pytesseract==0.3.13
PyMuPDF==1.24.11
requests==2.32.3
asyncpg